+++ details.js
@@ -1,5 +1,3 @@
-import cardsDetail1 from './data.js';
-
 const body = document.querySelector('body');
 const open = document.querySelectorAll('.open');
 
@@ -98,6 +96,90 @@ div6.setAttribute('class', 'footer');
 img2.addEventListener('click', () => {
   section.classList.remove('show');
 });
+
+const cardsDetail1 = [
+  {
+    name: 'Bar in Motion',
+    description: 'The Bar in Motion is a project done in the first Microverse module milestone. I created the project as a reference to the halloween holiday function organized by the Bar in Motion Communuty. The project was able to communicate the fun events that were organized for the day and also worked as a marketing strategy for the company by displaying their previous events',
+    canopy: 'CANOPY',
+    back: 'Back End Dev',
+    year: '2021',
+    image: 'images/bar1.png',
+    tech1: 'html',
+    tech2: 'css',
+    tech3: 'javascript',
+    closBtn: 'images/close.png',
+    buttn1: 'See live',
+    buttn2: 'See Source',
+    counterOne: 'images/Counter.jpg',
+    counterTwo: 'images/Counter.jpg',
+    live: 'images/live.jpg',
+    gity: 'images/gitty.jpg',
+    liveLink: 'https://mosams.github.io/Module-One-Capstone-Project/',
+    liveGit: 'https://github.com/Mosams/Module-One-Capstone-Project.git',
+  },
+
+  {
+    name: 'Multi-Post Stories',
+    description: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industrys standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent',
+    canopy: 'CANOPY',
+    back: 'Back End Dev',
+    year: '2015',
+    image: 'images/card2.jpg',
+    tech1: 'html',
+    tech2: 'css',
+    tech3: 'javascript',
+    closBtn: 'images/close.jpg',
+    buttn1: 'See live',
+    buttn2: 'See Source',
+    counterOne: 'images/Counter.jpg',
+    counterTwo: 'images/Counter.jpg',
+    live: 'images/live.jpg',
+    gity: 'images/gitty.jpg',
+    liveLink: 'https://mysterious-hollows-81066.herokuapp.com/',
+    liveGit: 'https://github.com/Mosams/Budget_Hippo.git',
+  },
+  {
+    name: 'Tonic',
+    description: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industrys standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent',
+    canopy: 'CANOPY',
+    back: 'Back End Dev',
+    year: '2015',
+    image: 'images/card3.jpg',
+    tech1: 'html',
+    tech2: 'css',
+    tech3: 'javascript',
+    closBtn: 'images/close.jpg',
+    buttn1: 'See live',
+    buttn2: 'See Source',
+    counterOne: 'images/Counter.jpg',
+    counterTwo: 'images/Counter.jpg',
+    live: 'images/live.jpg',
+    gity: 'images/gitty.jpg',
+    liveLink: 'https://mosams.github.io/Module-One-Capstone-Project/',
+    liveGit: 'https://github.com/Mosams/Module-One-Capstone-Project.git',
+  },
+  {
+    name: 'Multi-Post Stories',
+    description: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industrys standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent',
+    canopy: 'CANOPY',
+    back: 'Back End Dev',
+    year: '2015',
+    image: 'images/card4.jpg',
+    tech1: 'html',
+    tech2: 'css',
+    tech3: 'javascript',
+    closBtn: 'images/close.jpg',
+    buttn1: 'See live',
+    buttn2: 'See Source',
+    counterOne: 'images/Counter.jpg',
+    counterTwo: 'images/Counter.jpg',
+    live: 'images/live.jpg',
+    gity: 'images/gitty.jpg',
+    liveLink: 'https://mosams.github.io/Module-One-Capstone-Project/',
+    liveGit: 'https://github.com/Mosams/Module-One-Capstone-Project.git',
+  },
+];
 
 function displayInfo() {
   open.forEach((element) => {
