+++ index.html
@@ -261,8 +261,7 @@
         </div>
     </section>
     <script src="menu.js"></script>
-    <script type="module" src="data.js"></script>
-    <script type="module" src="details.js"></script>
+    <script src="details.js"></script>
     <script src="validation.js"></script>
 </body>
 </html>
