+++ package.json
@@ -2,8 +2,7 @@
   "name": "Personal-Portfolio",
   "version": "1.0.0",
   "description": "> This project involves creating a basic mobile version portfolio skeleton which will be used in the development of a complete personal portfolio accessible both through phone and desktop.",
-  "main": "menu.js",
-  "type": "module",
+  "main": "details.js",
   "devDependencies": {
     "babel-eslint": "^10.1.0",
     "eslint": "^7.32.0",
